+normalPlaceCards

+++ app/src/main/java/chiglintsev/notboringtrails20/models/Places.java
@@ -1,18 +1,14 @@
 package chiglintsev.notboringtrails20.models;
 
 
-import android.location.Location;
-
 import com.activeandroid.Model;
 import com.activeandroid.annotation.Column;
 import com.activeandroid.annotation.Table;
 
-import java.util.Comparator;
-
 @Table(name = "Places")
 public class Places extends Model {
 
-    public double distance;
+    public float distance;
 
 
     @Column(name = "name")
